Move metadata to it's own module
